refactor: update import paths to use .js extensions and adjust TypeScript configuration
