+++ src/routes/v1/index.ts
@@ -1,3 +1,3 @@
-export { default as buildUserRouter } from './user';
-export { default as buildGenerationRouter } from './generation';
-export { default as buildAuthRouter } from './auth';
+export { default as buildUserRouter } from './user.js';
+export { default as buildGenerationRouter } from './generation.js';
+export { default as buildAuthRouter } from './auth.js';
