+++ tsconfig.json
@@ -6,17 +6,20 @@
     "skipLibCheck": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "strict": true,
+    "strict": false,
     "forceConsistentCasingInFileNames": true,
-    "module": "commonjs",
+    "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
-    "isolatedModules": true,
+    "isolatedModules": false,
     "baseUrl": ".",
     "noFallthroughCasesInSwitch": true,
-    "outDir": "./backend"
+    "outDir": "./backend",
+    "noImplicitAny": false,
+    "noImplicitReturns": false,
+    "noImplicitThis": false
   },
   "compileOnSave": true,
-  "include": ["./src/**/*", "./package.json"],
-
+  "include": ["./src/**/*"],
+  "exclude": ["node_modules", "backend", "dist"]
 }
